+++ cpp/perspective/src/cpp/view.cpp
@@ -9,7 +9,7 @@
 
 #include <perspective/first.h>
 #include <perspective/view.h>
-#include <perspective/arrow.h>
+#include <perspective/arrow_writer.h>
 #include <sstream>
